Reset dualstack metrics before exporting, move them to separate file
